// src/app/core/enum/visit-status.enum.ts
export enum VisitStatus {
  SCHEDULE = 'schedule',
  UNCONFIRMED = 'unconfirmed',
  CONFIRMED = 'confirmed',
  RESCHEDULE = 'reschedule',
  ABSENT = 'absent',
  MISSED = 'missed',
  CLINIC = 'clinic',
  REVIEW = 'review',
  FINISHED = 'finished',
  VALIDATED = 'validated',
  ENDORSED = 'endorsed',
  MONITOR = 'monitor',
}